requirements_manual.txt
psycopg2-binary
requests==2.28.1